pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "china-shift-share-dummy"
version = "0.0.1"
description = "Dummy pipeline scaffold for China shift-share analysis"
authors = [{name = "Factory Droid"}]
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0"
]

[tool.pytest.ini_options]
pythonpath = ["scripts"]
